Refactoring Reload when back from detail to the favorite view(only when data is dirty) Add RecyclerView Header and fix some bugs Remove leanback library

+++ PopularMovies-ST2/app/src/main/java/com/example/popularmovies/model/Detail.java
@@ -26,7 +26,7 @@ public class Detail {
     List<ProductionCountry> productionCountries;
     @SerializedName("release_date")
     String releaseDate;
-    Integer revenue;
+    Long revenue;
     Integer runtime;
     List<SpokenLaunguage> spokenLanguages;
     String status;
@@ -152,11 +152,11 @@ public class Detail {
         this.releaseDate = releaseDate;
     }
 
-    public Integer getRevenue() {
+    public Long getRevenue() {
         return revenue;
     }
 
-    public void setRevenue(Integer revenue) {
+    public void setRevenue(Long revenue) {
         this.revenue = revenue;
     }
